Block: move block and header into directories
