+++ packages/block/src/index.ts
@@ -1,8 +1,8 @@
-export { Block } from './block.js'
+export { Block } from './block/block.js'
+export * from './block/index.js'
 export * from './consensus/index.js'
-export * from './constructors.js'
 export { type BeaconPayloadJson, executionPayloadFromBeaconPayload } from './from-beacon-payload.js'
-export { BlockHeader } from './header.js'
+export * from './header/index.js'
 export {
   genRequestsTrieRoot,
   genTransactionsTrieRoot,
